Fixed signup errors causing crashes
Refactored gorouter implementation.
